[FEAT] 유저 정보 전역관리 fetch하는 hook 추가 #55
